{{cookiecutter.project_name}}/requirements.txt
Django~=1.9
SQLAlchemy==0.9.8
dj-database-url==0.3.0
django-postgrespool==0.3.0
djangorestframework==3.3
gunicorn==19.1.1
psycopg2==2.5.4
whitenoise==1.0.3
wsgiref==0.1.2
